+++ packages/react-core/src/components/Dropdown/Toggle.tsx
@@ -109,7 +109,7 @@ export class Toggle extends React.Component<ToggleProps> {
     event.preventDefault();
     if ((event.key === 'Tab' || event.key === 'Enter' || event.key === ' ') && this.props.isOpen) {
       this.props.onToggle(!this.props.isOpen, event);
-    } else if ((event.key === 'Enter' || event.key === ' ') && !this.props.isOpen) {
+    } else if ((event.key === 'Enter' || event.key === ' ' || event.key === 'ArrowDown') && !this.props.isOpen) {
       this.props.onToggle(!this.props.isOpen, event);
       this.props.onEnter();
     }