+++ packages/react-core/src/components/Nav/NavExpandable.tsx
@@ -95,23 +95,22 @@ export class NavExpandable extends React.Component<NavExpandableProps, NavExpand
 
   render() {
     const {
-      /* eslint-disable @typescript-eslint/no-unused-vars */
-      id,
-      groupId,
-      isExpanded,
-      onExpand,
-      ouiaId,
-      ouiaSafe,
-      /* eslint-enable @typescript-eslint/no-unused-vars */
       title,
       srText,
       children,
       className,
       isActive,
+      ouiaId,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      groupId,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      id,
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      isExpanded,
       ...props
     } = this.props;
 
-    const { expandedState } = this.state;
+    const { expandedState, ouiaStateId } = this.state;
 
     const onClick = () => {
       this.setState(prevState => ({ expandedState: !prevState.expandedState }));
@@ -128,10 +127,7 @@ export class NavExpandable extends React.Component<NavExpandableProps, NavExpand
               isActive && styles.modifiers.current,
               className
             )}
-            {...getOUIAProps(
-              NavExpandable.displayName,
-              this.props.ouiaId !== undefined ? this.props.ouiaId : this.state.ouiaStateId
-            )}
+            {...getOUIAProps(NavExpandable.displayName, ouiaId !== undefined ? ouiaId : ouiaStateId)}
             onClick={(e: React.MouseEvent<HTMLLIElement, MouseEvent>) => this.handleToggle(e, context.onToggle)}
             {...props}
           >