+++ packages/react-datetime/src/components/DatePicker/DatePicker.tsx
@@ -38,6 +38,8 @@ export interface DatePickerProps
   beforeMinDateErrorMessage?: string;
   /** Error message to display when the date is provided after the provided maxDate. */
   afterEndDateErrorMessage?: string;
+  /** Callback called every time the input value changes */
+  onChange?: (value: string, date?: Date) => void;
 }
 
 interface DatePickerState {
@@ -84,7 +86,8 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
     locale: DatePickerLocales.en,
     placeholder: 'mm/dd/yyyy',
     value: '',
-    'aria-label': 'Date picker'
+    'aria-label': 'Date picker',
+    onChange: (): any => undefined
   };
 
   constructor(props: DatePickerProps) {
@@ -101,7 +104,7 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
   private calendar: any = null;
 
   componentDidMount() {
-    const { dateFormat, minDate, maxDate, locale, value } = this.props;
+    const { dateFormat, minDate, maxDate, locale, value, onChange } = this.props;
     const { flatpickrDateFormat } = this.state;
 
     this.calendar = flatpickr(this.inputEl.current, {
@@ -113,8 +116,12 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
       allowInput: true,
       static: true,
       allowInvalidPreload: true,
+      onChange: (dates, currentDateString) => {
+        this.validateDate(currentDateString);
+        onChange(currentDateString, dates[0]);
+      },
       errorHandler: () => {
-        this.handleError(this.props.invalidFormatErrorMessage || `Please use format ${dateFormat}.`);
+        this.handleError(this.props.invalidFormatErrorMessage || `Please use format: ${dateFormat}.`);
       },
       onReady: (date, format, fp) => {
         fp.calendarContainer.classList.add(styles.datePickerCalendar);
@@ -143,7 +150,7 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
     this.calendar && this.calendar.isOpen ? this.calendar.open() : this.calendar.close();
   };
 
-  validateDate = (dateStr: string, event: React.KeyboardEvent | React.ChangeEvent) => {
+  validateDate = (dateStr: string, event?: React.KeyboardEvent | React.ChangeEvent) => {
     const {
       dateFormat,
       minDate: minDateIn,
@@ -156,7 +163,9 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
 
     const { flatpickrDateFormat } = this.state;
     if (dateStr === '') {
-      event.stopPropagation();
+      if (event) {
+        event.stopPropagation();
+      }
       return;
     }
 
@@ -165,11 +174,15 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
     const maxDate = this.calendar.parseDate(maxDateIn, flatpickrDateFormat);
     if (!isValid(date)) {
       this.handleError(invalidFormatErrorMessage || `Please use format: ${dateFormat}.`);
-      event.stopPropagation();
+      if (event) {
+        event.stopPropagation();
+      }
     } else if (isValid(minDate) && isValid(maxDate)) {
       if (date < minDate || date > maxDate) {
         this.handleError(dateOutOfRangeErrorMessage || `The date is outside the allowable range.`);
-        event.stopPropagation();
+        if (event) {
+          event.stopPropagation();
+        }
       } else {
         this.setState({
           invalid: false,
@@ -178,10 +191,14 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
       }
     } else if (isValid(minDate) && date < minDate) {
       this.handleError(beforeMinDateErrorMessage || `Date is before the allowable range.`);
-      event.stopPropagation();
+      if (event) {
+        event.stopPropagation();
+      }
     } else if (isValid(maxDate) && date > maxDate) {
       this.handleError(afterEndDateErrorMessage || `Date is after the allowable range.`);
-      event.stopPropagation();
+      if (event) {
+        event.stopPropagation();
+      }
     } else {
       this.setState({
         invalid: false,
@@ -195,7 +212,6 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
   };
 
   handleError = (errorMessage: string) => {
-    this.calendar.close();
     this.setState({
       invalid: true,
       invalidText: errorMessage
@@ -221,6 +237,7 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
       beforeMinDateErrorMessage,
       afterEndDateErrorMessage,
       /* eslint-enable @typescript-eslint/no-unused-vars */
+      onChange,
       ...props
     } = this.props;
 
@@ -229,8 +246,9 @@ export class DatePicker extends React.Component<DatePickerProps, DatePickerState
         <TextInput
           isDisabled={isDisabled}
           iconVariant="calendar"
-          onClick={this.toggleCalendar}
+          onFocus={this.toggleCalendar}
           onBlur={this.onBlur}
+          onChange={dateStr => onChange(dateStr)}
           ref={this.inputEl}
           aria-label={ariaLabel}
           placeholder={placeholder}