+++ packages/react-integration/cypress/integration/select.spec.ts
@@ -5,6 +5,37 @@ describe('Select Test', () => {
     cy.url().should('eq', 'http://localhost:3000/select-demo-nav-link');
   });
 
+  it('Verify Select with Divider', () => {
+    cy.get('#single-select-with-divider').click();
+    cy.get('#Miss > .pf-c-select__menu-item').click();
+    cy.get('#single-select-with-divider')
+      .contains('Miss')
+      .should('exist');
+
+    cy.get('#single-select-with-divider').click();
+    cy.get('#Mr > .pf-c-select__menu-item').click();
+    cy.get('#single-select-with-divider')
+      .contains('Mr')
+      .should('exist');
+  });
+
+  it('Verify Typeahead Select with labels', () => {
+    const find = (selector: string) =>
+      cy
+        .get('#typeahead-select-labels-id')
+        .parent()
+        .find(selector);
+    find('button.pf-c-select__toggle-button').click();
+    find('li:nth-child(3) button').click();
+    find('#typeahead-select-label-select-typeahead').should('have.value', 'Florida');
+    find('button.pf-c-select__toggle-clear:first').click();
+    find('#typeahead-select-label-select-typeahead').should('have.value', '');
+    cy.get('#button-typeahead-labels').click();
+    find('#typeahead-select-label-select-typeahead').should('have.value', 'New York');
+    find('button.pf-c-select__toggle-clear:first').click();
+    find('#typeahead-select-label-select-typeahead').should('have.value', '');
+  });
+
   it('Verify Single Select', () => {
     cy.get('#single-select').click();
     cy.get('#Miss > .pf-c-select__menu-item').click();