+++ packages/react-catalog-view-extension/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.75](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.74...@patternfly/react-catalog-view-extension@4.8.75) (2020-10-21)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.74](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.73...@patternfly/react-catalog-view-extension@4.8.74) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.73](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.72...@patternfly/react-catalog-view-extension@4.8.73) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.72](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.71...@patternfly/react-catalog-view-extension@4.8.72) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.71](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.70...@patternfly/react-catalog-view-extension@4.8.71) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.70](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.69...@patternfly/react-catalog-view-extension@4.8.70) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [4.8.69](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.68...@patternfly/react-catalog-view-extension@4.8.69) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
 ## [4.8.68](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@4.8.67...@patternfly/react-catalog-view-extension@4.8.68) (2020-10-13)
 
 **Note:** Version bump only for package @patternfly/react-catalog-view-extension