+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.35](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.34...@patternfly/react-docs@5.10.35) (2020-10-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.34](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.33...@patternfly/react-docs@5.10.34) (2020-10-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.32...@patternfly/react-docs@5.10.33) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.31...@patternfly/react-docs@5.10.32) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.30...@patternfly/react-docs@5.10.31) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.29...@patternfly/react-docs@5.10.30) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.28...@patternfly/react-docs@5.10.29) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.27...@patternfly/react-docs@5.10.28) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.26...@patternfly/react-docs@5.10.27) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.10.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.25...@patternfly/react-docs@5.10.26) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [5.10.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.10.24...@patternfly/react-docs@5.10.25) (2020-10-13)
 
 **Note:** Version bump only for package @patternfly/react-docs