+++ packages/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.133](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.132...@patternfly/react-inline-edit-extension@4.5.133) (2020-10-21)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.132](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.131...@patternfly/react-inline-edit-extension@4.5.132) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.131](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.130...@patternfly/react-inline-edit-extension@4.5.131) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.130](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.129...@patternfly/react-inline-edit-extension@4.5.130) (2020-10-20)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.129](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.128...@patternfly/react-inline-edit-extension@4.5.129) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.128](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.127...@patternfly/react-inline-edit-extension@4.5.128) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.127](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.126...@patternfly/react-inline-edit-extension@4.5.127) (2020-10-19)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.126](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.125...@patternfly/react-inline-edit-extension@4.5.126) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.5.125](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.124...@patternfly/react-inline-edit-extension@4.5.125) (2020-10-14)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [4.5.124](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.5.123...@patternfly/react-inline-edit-extension@4.5.124) (2020-10-13)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension